Exp. 9.1. Different output values
This is a simple experiment to test whether a small SNN can learn to
output two different values on the output neurons (something that my
previous versions struggled with). Apparently, there was a bug in the
fitness.py where the array of output nodes' values was not addressed
properly (i.e., individual neurons should have been accesssed but
instead the whole matrix was processed altogether, which lead to output
neurons always firing very similarly).