Exp. 9.3
Exp. 9.2. was skipped because it only consisted of adding an option to
turn off intra-layer connections within the hidden layer (thought to be
the reason for output neurons never learning to output different values)
Exp. 9.3. Focused on trainng the SNN to exhibit one set of values for
the first half of the simulation and then switch to a different set of
output values. The switch is governed by a pair of so-called "cortical"
neurons. Cortical neuron 1 feeds 1.0 into the hidden layer while
cortical neuron 2 feeds 0.0 for the first half, then they switch.