Exp. 9.4.
Exp. 9.4 is an extension of 9.3: the SNN is trained to exhibit oen set
of output values then to switch to another set of output values, and yet
again to switch back to the original set. The purpose was to test the
limits of memory of a quite small SNN that consisted only of 5 hidden
spiking neurons.