Exp 9.7a Proof-of-concept
The basal ganglia (BG) are added as external nodes with hardcoded firing
patterns and are connected to the hidden layer via evolvable weights.
The goal of this experiment was to show that this morphology can be
optimized to produce both swithcing when the BG firing is desynchronized
(firing is a gaussian process) and resist switching when the BG firing
is synchronized.
The fitness function chosen for this experiment ran each population
member twice (when the BG were SYNC and DESYNC) and assigned fitness
scores to both runs according to the goal (to switch when the BG were
DESYNC and to resist switching when the BG were SYNC). The lowest of two
scores was selected as the final score (the higher the score the better
the behavior). This score selection is rather strict and may prohibit
successful evolution. Other options: average scores, sum scores,
multiply scores, or try multi-objective evolution (e.g., AFPO).