EXP 9.7b Hypothesis testing
Exp. 9.7b is the extension of the exp 9.7a (proof-of-concept). Now,
instead of explicitly evolving the network for exhibiting the desired
behavior (switch only when the basal ganglia are DESYNC), this
experiment aims at evolving networks that switch between states and the
BG briefly DESYNC just before and slightly after (-50 ticks, + 50 ticks)
the SM loop receives switching signal from CTX neurons. Successfully
trained networks will have to undergo several tests after training:
* Does the network achieve switching when the time for switching changes
(to check if the results were a consequence of overfitting)
* Does the network still switch when the BG stop DESYNCing? The
hypothesis is that the switching will be impaired.
* If CTX is removed, does BG DESYNC change the output of the SM loop
permanently (i.e., does the BG drive the change?). It's expected that
once the CTX is removed, there is no change regardless of SYNC/DESYNC in
the BG